chore: move payment related pages into Payments folder
